data_quality_monitoring/requirements.txt
apscheduler>=3.10.0
pulsar-client>=3.5.0
pymongo>=4.6.3
pymysql>=1.1.1
loguru>=0.7.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=24.0.0
isort>=5.13.0
